gen_ai_practical_exercise_4/requirements.txt
fastapi
uvicorn
langchain
chromadb
ollama
google-genai
python-dotenv
